Create course fee.go
The students are entitled to get scholarship based on the marks scored in the qualifying exam. Scholarship percentage should be considered as half of the marks scored by the student on the course fee. Apart from the course fee (after deducting the scholarship amount), the students have to pay an extra service fees.